fix: drop a same-type cast that `-D warnings` rejects
`check-test-targets` was red on main, which fails `ci-matrix` (and `ci`) before
any test runs:

    error: casting to the same type is unnecessary (`u32` -> `u32`)
      --> packages/testing/nros-tests/src/lib.rs:967:21

`TEST_DOMAIN_MAX` and `DOMAINS_PER_SLOT` are both `const … : u32`, so
`(TEST_DOMAIN_MAX / DOMAINS_PER_SLOT) as u32` casts u32 to u32. From
`8458bec01` (#703 follow-up), where the ceiling moved and the cast was carried
along.

Caught by tier 2 rather than by tier 1's fast line: `check-test-targets` lives
in `check-build`, not `check-fast`.

+++ packages/testing/nros-tests/src/lib.rs
@@ -964,7 +964,7 @@ mod tests {
     /// fails on 14 pairs.
     #[test]
     fn a_slots_domains_never_land_on_a_live_neighbours() {
-        let slots = (TEST_DOMAIN_MAX / DOMAINS_PER_SLOT) as u32;
+        let slots = TEST_DOMAIN_MAX / DOMAINS_PER_SLOT;
         let mut owner = std::collections::HashMap::new();
         for slot in 0..slots {
             for seq in 0..DOMAINS_PER_SLOT {
